revert: restore all listed building test files to origin/main state
