+++ odw/test/integration_test/etl/curated/test_listed_building_curated_process.py
@@ -1,13 +1,37 @@
 import mock
 import pytest
 import odw.test.util.mock.import_mock_notebook_utils  # noqa: F401
-from pyspark.sql import functions as F
 from pyspark.sql.types import LongType, StringType, StructField, StructType
 from odw.core.etl.transformation.curated.listed_building_curated_process import ListedBuildingCuratedProcess
 from odw.test.integration_test.etl.etl_test_case import ETLTestCase
 from odw.test.util.session_util import PytestSparkSessionUtil
+from odw.test.util.assertion import assert_dataframes_equal, assert_etl_result_successful
 
 pytestmark = pytest.mark.xfail(reason="Curated logic not implemented yet")
+
+
+def _harmonised_row(**overrides):
+    base = {
+        "dataset": "listed-building",
+        "endDate": None,
+        "entity": "1001",
+        "entryDate": "2024-01-01",
+        "geometry": "POLYGON((1 1,2 2,3 3,1 1))",
+        "listedBuildingGrade": "II",
+        "name": "Building One",
+        "organisationEntity": "org-1",
+        "point": "POINT(1 1)",
+        "prefix": "listed-building",
+        "reference": "LB-001",
+        "startDate": "2020-01-01",
+        "typology": "grade-ii",
+        "documentationUrl": "https://example.com/lb-001",
+        "dateReceived": "2025-01-01",
+        "rowID": "row-1",
+        "validTo": "None",
+        "isActive": "Y",
+    }
+    return base | overrides
 
 
 def _harmonised_schema():
@@ -46,358 +70,64 @@ def _curated_schema():
     )
 
 
-def _harmonised_df(spark):
-    return spark.createDataFrame(
-        [
-            (
-                "listed-building",
-                None,
-                "1001",
-                "2024-01-01",
-                "POLYGON((1 1,2 2,3 3,1 1))",
-                "II",
-                "Building One",
-                "org-1",
-                "POINT(1 1)",
-                "listed-building",
-                "LB-001",
-                "2020-01-01",
-                "grade-ii",
-                "https://example.com/lb-001",
-                "2025-01-01",
-                "row-1",
-                None,
-                "Y",
-            ),
-            (
-                "listed-building",
-                None,
-                "1002",
-                "2024-02-01",
-                None,
-                "I",
-                "Building Two",
-                "org-2",
-                None,
-                "listed-building",
-                "LB-002",
-                "2021-01-01",
-                "grade-i",
-                None,
-                "2025-01-01",
-                "row-2",
-                None,
-                "Y",
-            ),
-            (
-                "listed-building",
-                None,
-                "1003",
-                "2024-03-01",
-                None,
-                "II*",
-                "Inactive Building",
-                "org-3",
-                None,
-                "listed-building",
-                "LB-003",
-                "2022-01-01",
-                "grade-ii-star",
-                None,
-                "2025-01-01",
-                "row-3",
-                None,
-                "N",
-            ),
-        ],
-        schema=_harmonised_schema(),
-    )
-
-
-def _changed_harmonised_df(spark):
-    return spark.createDataFrame(
-        [
-            (
-                "listed-building",
-                None,
-                "1001",
-                "2024-01-01",
-                "POLYGON((1 1,2 2,3 3,1 1))",
-                "II",
-                "Building One Updated",
-                "org-1",
-                "POINT(1 1)",
-                "listed-building",
-                "LB-001",
-                "2020-01-01",
-                "grade-ii",
-                "https://example.com/lb-001",
-                "2025-01-01",
-                "row-9",
-                None,
-                "Y",
-            ),
-        ],
-        schema=_harmonised_schema(),
-    )
-
-
-def _duplicate_harmonised_df(spark):
-    return spark.createDataFrame(
-        [
-            (
-                "listed-building",
-                None,
-                "1001",
-                "2024-01-01",
-                "POLYGON((1 1,2 2,3 3,1 1))",
-                "II",
-                "Building One",
-                "org-1",
-                "POINT(1 1)",
-                "listed-building",
-                "LB-001",
-                "2020-01-01",
-                "grade-ii",
-                "https://example.com/lb-001",
-                "2025-01-01",
-                "row-1",
-                None,
-                "Y",
-            ),
-            (
-                "listed-building",
-                None,
-                "1001",
-                "2024-01-01",
-                "POLYGON((1 1,2 2,3 3,1 1))",
-                "II",
-                "Building One",
-                "org-1",
-                "POINT(1 1)",
-                "listed-building",
-                "LB-001",
-                "2020-01-01",
-                "grade-ii",
-                "https://example.com/lb-001",
-                "2025-01-01",
-                "row-2",
-                None,
-                "Y",
-            ),
-        ],
-        schema=_harmonised_schema(),
-    )
-
-
-def _empty_harmonised_df(spark):
-    return spark.createDataFrame([], schema=_harmonised_schema())
-
-
-def _existing_curated_df(spark):
-    return spark.createDataFrame(
-        [
-            (1001, "LB-001", "Building One", "II"),
-        ],
-        schema=_curated_schema(),
-    )
-
-
 class TestListedBuildingCuratedProcess(ETLTestCase):
-    def test__listed_building_curated_process__run__initial_load_matches_legacy(self):
+    def test__listed_building_curated_process__run__with_no_existing_data(self):
         spark = PytestSparkSessionUtil().get_spark_session()
-
-        source_data = {
-            "source_data": _harmonised_df(spark),
-            "target_exists": False,
-        }
-
-        with (
-            mock.patch("odw.core.etl.etl_process.LoggingUtil") as mock_etl_logging,
-            mock.patch("odw.core.etl.transformation.curated.listed_building_curated_process.LoggingUtil") as mock_process_logging,
-        ):
-            mock_etl_logging.return_value = mock.Mock()
-            mock_process_logging.return_value = mock.Mock()
-
-            inst = ListedBuildingCuratedProcess(spark)
-
-            with (
-                mock.patch.object(inst, "load_data", return_value=source_data),
-                mock.patch.object(inst, "write_data") as mock_write,
-            ):
-                result = inst.run()
-
-        data_to_write = mock_write.call_args[0][0]
-        df = data_to_write[inst.OUTPUT_TABLE]["data"]
-
-        assert data_to_write[inst.OUTPUT_TABLE]["write_mode"] == "overwrite"
-        assert df.count() == 2
-        assert result.metadata.insert_count == 2
-        assert result.metadata.update_count == 0
-
-        assert df.columns == ["entity", "reference", "name", "listedBuildingGrade"]
-
-        entity_ids = {row["entity"] for row in df.select("entity").collect()}
-        assert entity_ids == {1001, 1002}
-        assert df.where(F.col("entity") == 1003).count() == 0
-
-    def test__listed_building_curated_process__run__updates_existing_entity_when_non_key_fields_change(
-        self,
-    ):
-        spark = PytestSparkSessionUtil().get_spark_session()
-
-        source_data = {
-            "source_data": _changed_harmonised_df(spark),
-            "target_data": _existing_curated_df(spark),
-            "target_exists": True,
-        }
-
-        with (
-            mock.patch("odw.core.etl.etl_process.LoggingUtil") as mock_etl_logging,
-            mock.patch("odw.core.etl.transformation.curated.listed_building_curated_process.LoggingUtil") as mock_process_logging,
-        ):
-            mock_etl_logging.return_value = mock.Mock()
-            mock_process_logging.return_value = mock.Mock()
-
-            inst = ListedBuildingCuratedProcess(spark)
-
-            with (
-                mock.patch.object(inst, "load_data", return_value=source_data),
-                mock.patch.object(inst, "write_data") as mock_write,
-            ):
-                result = inst.run()
-
-        data_to_write = mock_write.call_args[0][0]
-        df = data_to_write[inst.OUTPUT_TABLE]["data"]
-
-        row = df.where(F.col("entity") == 1001).collect()[0]
-
-        assert row["name"] == "Building One Updated"
-        assert result.metadata.insert_count == 0
-        assert result.metadata.update_count == 1
-
-    def test__listed_building_curated_process__run__does_not_duplicate_identical_existing_entity(
-        self,
-    ):
-        spark = PytestSparkSessionUtil().get_spark_session()
-
-        identical_source = spark.createDataFrame(
-            [
-                (
-                    "listed-building",
-                    None,
-                    "1001",
-                    "2024-01-01",
-                    "POLYGON((1 1,2 2,3 3,1 1))",
-                    "II",
-                    "Building One",
-                    "org-1",
-                    "POINT(1 1)",
-                    "listed-building",
-                    "LB-001",
-                    "2020-01-01",
-                    "grade-ii",
-                    "https://example.com/lb-001",
-                    "2025-01-01",
-                    "row-1",
-                    None,
-                    "Y",
-                ),
-            ],
+        test_case = "t_lbcpr_wned"
+        harmonised_data = spark.createDataFrame(
+            (
+                _harmonised_row(name="Building One", entity="1001", reference="LB-001", listedBuildingGrade="I"),
+                _harmonised_row(name="Building Two", entity="1002", reference="LB-002", listedBuildingGrade="I"),
+            ),
             schema=_harmonised_schema(),
         )
-
-        source_data = {
-            "source_data": identical_source,
-            "target_data": _existing_curated_df(spark),
-            "target_exists": True,
-        }
-
-        with (
-            mock.patch("odw.core.etl.etl_process.LoggingUtil") as mock_etl_logging,
-            mock.patch("odw.core.etl.transformation.curated.listed_building_curated_process.LoggingUtil") as mock_process_logging,
-        ):
-            mock_etl_logging.return_value = mock.Mock()
-            mock_process_logging.return_value = mock.Mock()
-
+        table_name = f"{test_case}_listed_building"
+        self.write_existing_table(spark, harmonised_data, table_name, "odw_harmonised_db", "odw-harmonised", table_name, "overwrite")
+        expected_curated_data_after_writing = spark.createDataFrame(
+            (
+                {"entity": 1001, "reference": "LB-001", "name": "Building One", "listedBuildingGrade": "I"},
+                {"entity": 1002, "reference": "LB-002", "name": "Building Two", "listedBuildingGrade": "I"},
+            ),
+            schema=_curated_schema(),
+        )
+        with mock.patch.object(ListedBuildingCuratedProcess, "OUTPUT_TABLE", table_name):
             inst = ListedBuildingCuratedProcess(spark)
+            result = inst.run()
+            assert_etl_result_successful(result)
+            actual_table_data = spark.table(f"odw_curated_db.{table_name}")
+            assert_dataframes_equal(expected_curated_data_after_writing, actual_table_data)
 
-            with (
-                mock.patch.object(inst, "load_data", return_value=source_data),
-                mock.patch.object(inst, "write_data") as mock_write,
-            ):
-                result = inst.run()
-
-        data_to_write = mock_write.call_args[0][0]
-        df = data_to_write[inst.OUTPUT_TABLE]["data"]
-
-        assert df.count() == 1
-        assert df.where(F.col("entity") == 1001).count() == 1
-        assert result.metadata.insert_count == 0
-        assert result.metadata.update_count == 0
-
-    def test__listed_building_curated_process__run__drops_duplicate_source_rows(
-        self,
-    ):
+    def test__listed_building_curated_process__run__with_existing_data(self):
         spark = PytestSparkSessionUtil().get_spark_session()
-
-        source_data = {
-            "source_data": _duplicate_harmonised_df(spark),
-            "target_exists": False,
-        }
-
-        with (
-            mock.patch("odw.core.etl.etl_process.LoggingUtil") as mock_etl_logging,
-            mock.patch("odw.core.etl.transformation.curated.listed_building_curated_process.LoggingUtil") as mock_process_logging,
-        ):
-            mock_etl_logging.return_value = mock.Mock()
-            mock_process_logging.return_value = mock.Mock()
-
+        test_case = "t_lbcpr_wed"
+        harmonised_data = spark.createDataFrame(
+            (
+                _harmonised_row(
+                    name="Building One Renamed", entity="1001", reference="LB-001 Updated", listedBuildingGrade="I Updated"
+                ),  # Should be updated
+                _harmonised_row(name="Building Two", entity="1002", reference="LB-002", listedBuildingGrade="I"),
+            ),
+            schema=_harmonised_schema(),
+        )
+        table_name = f"{test_case}_listed_building"
+        self.write_existing_table(spark, harmonised_data, table_name, "odw_harmonised_db", "odw-harmonised", table_name, "overwrite")
+        curated_data = spark.createDataFrame(
+            (
+                {"entity": 1001, "reference": "LB-001", "name": "Building One", "listedBuildingGrade": "I"},  # Should be updated
+            ),
+            schema=_curated_schema(),
+        )
+        self.write_existing_table(spark, curated_data, table_name, "odw_curated_db", "odw-curated", table_name, "overwrite")
+        expected_curated_data_after_writing = spark.createDataFrame(
+            (
+                {"entity": 1001, "reference": "LB-001 Updated", "name": "Building One Renamed", "listedBuildingGrade": "I Updated"},  # Updated
+                {"entity": 1002, "reference": "LB-002", "name": "Building Two", "listedBuildingGrade": "I"},
+            ),
+            schema=_curated_schema(),
+        )
+        with mock.patch.object(ListedBuildingCuratedProcess, "OUTPUT_TABLE", table_name):
             inst = ListedBuildingCuratedProcess(spark)
-
-            with (
-                mock.patch.object(inst, "load_data", return_value=source_data),
-                mock.patch.object(inst, "write_data") as mock_write,
-            ):
-                result = inst.run()
-
-        data_to_write = mock_write.call_args[0][0]
-        df = data_to_write[inst.OUTPUT_TABLE]["data"]
-
-        assert df.count() == 1
-        assert df.where(F.col("entity") == 1001).count() == 1
-        assert result.metadata.insert_count == 1
-        assert result.metadata.update_count == 0
-
-    def test__listed_building_curated_process__run__empty_source_returns_empty_output(
-        self,
-    ):
-        spark = PytestSparkSessionUtil().get_spark_session()
-
-        source_data = {
-            "source_data": _empty_harmonised_df(spark),
-            "target_exists": False,
-        }
-
-        with (
-            mock.patch("odw.core.etl.etl_process.LoggingUtil") as mock_etl_logging,
-            mock.patch("odw.core.etl.transformation.curated.listed_building_curated_process.LoggingUtil") as mock_process_logging,
-        ):
-            mock_etl_logging.return_value = mock.Mock()
-            mock_process_logging.return_value = mock.Mock()
-
-            inst = ListedBuildingCuratedProcess(spark)
-
-            with (
-                mock.patch.object(inst, "load_data", return_value=source_data),
-                mock.patch.object(inst, "write_data") as mock_write,
-            ):
-                result = inst.run()
-
-        data_to_write = mock_write.call_args[0][0]
-        df = data_to_write[inst.OUTPUT_TABLE]["data"]
-
-        assert data_to_write[inst.OUTPUT_TABLE]["write_mode"] == "overwrite"
-        assert df.count() == 0
-        assert result.metadata.insert_count == 0
-        assert result.metadata.update_count == 0
+            result = inst.run()
+            assert_etl_result_successful(result)
+            actual_table_data = spark.table(f"odw_curated_db.{table_name}")
+            assert_dataframes_equal(expected_curated_data_after_writing, actual_table_data)
