+++ odw/test/unit_test/etl/standardised/test_listed_building_standardisation_process.py
@@ -405,9 +405,8 @@ class TestListedBuildingStandardisationProcess(SparkTestCase):
             "listed_building_outline_data": _listed_building_outline_raw_df(spark),
         }
 
-        with mock.patch("odw.core.etl.transformation.standardised.listed_building_standardisation_process.LoggingUtil"):
-            inst = ListedBuildingStandardisationProcess(spark)
-            data_to_write, _ = inst.process(source_data=source_data)
+        inst = ListedBuildingStandardisationProcess(spark)
+        data_to_write, _ = inst.process(source_data=source_data)
 
         assert LISTED_BUILDING_OUTPUT_TABLE in data_to_write
         assert LISTED_BUILDING_OUTLINE_OUTPUT_TABLE in data_to_write
@@ -422,9 +421,8 @@ class TestListedBuildingStandardisationProcess(SparkTestCase):
             "listed_building_outline_data": _listed_building_outline_raw_df(spark),
         }
 
-        with mock.patch("odw.core.etl.transformation.standardised.listed_building_standardisation_process.LoggingUtil"):
-            inst = ListedBuildingStandardisationProcess(spark)
-            data_to_write, _ = inst.process(source_data=source_data)
+        inst = ListedBuildingStandardisationProcess(spark)
+        data_to_write, _ = inst.process(source_data=source_data)
 
         df = data_to_write[LISTED_BUILDING_OUTPUT_TABLE]["data"]
 
@@ -474,9 +472,8 @@ class TestListedBuildingStandardisationProcess(SparkTestCase):
             "listed_building_outline_data": _listed_building_outline_raw_df(spark),
         }
 
-        with mock.patch("odw.core.etl.transformation.standardised.listed_building_standardisation_process.LoggingUtil"):
-            inst = ListedBuildingStandardisationProcess(spark)
-            data_to_write, _ = inst.process(source_data=source_data)
+        inst = ListedBuildingStandardisationProcess(spark)
+        data_to_write, _ = inst.process(source_data=source_data)
 
         df = data_to_write[LISTED_BUILDING_OUTLINE_OUTPUT_TABLE]["data"]
 
@@ -534,9 +531,8 @@ class TestListedBuildingStandardisationProcess(SparkTestCase):
             "listed_building_outline_data": _listed_building_outline_raw_df(spark),
         }
 
-        with mock.patch("odw.core.etl.transformation.standardised.listed_building_standardisation_process.LoggingUtil"):
-            inst = ListedBuildingStandardisationProcess(spark)
-            data_to_write, _ = inst.process(source_data=source_data)
+        inst = ListedBuildingStandardisationProcess(spark)
+        data_to_write, _ = inst.process(source_data=source_data)
 
         listed_building_df = data_to_write[LISTED_BUILDING_OUTPUT_TABLE]["data"]
         listed_building_outline_df = data_to_write[LISTED_BUILDING_OUTLINE_OUTPUT_TABLE]["data"]
@@ -579,9 +575,8 @@ class TestListedBuildingStandardisationProcess(SparkTestCase):
             "listed_building_outline_data": _listed_building_outline_raw_df(spark),
         }
 
-        with mock.patch("odw.core.etl.transformation.standardised.listed_building_standardisation_process.LoggingUtil"):
-            inst = ListedBuildingStandardisationProcess(spark)
-            data_to_write, _ = inst.process(source_data=source_data)
+        inst = ListedBuildingStandardisationProcess(spark)
+        data_to_write, _ = inst.process(source_data=source_data)
 
         assert data_to_write[LISTED_BUILDING_OUTPUT_TABLE]["write_mode"] == "overwrite"
         assert data_to_write[LISTED_BUILDING_OUTLINE_OUTPUT_TABLE]["write_mode"] == "overwrite"
@@ -596,9 +591,8 @@ class TestListedBuildingStandardisationProcess(SparkTestCase):
             "listed_building_outline_data": _listed_building_outline_raw_df(spark),
         }
 
-        with mock.patch("odw.core.etl.transformation.standardised.listed_building_standardisation_process.LoggingUtil"):
-            inst = ListedBuildingStandardisationProcess(spark)
-            data_to_write, _ = inst.process(source_data=source_data)
+        inst = ListedBuildingStandardisationProcess(spark)
+        data_to_write, _ = inst.process(source_data=source_data)
 
         listed_building_df = data_to_write[LISTED_BUILDING_OUTPUT_TABLE]["data"]
         listed_building_outline_df = data_to_write[LISTED_BUILDING_OUTLINE_OUTPUT_TABLE]["data"]
@@ -651,9 +645,8 @@ class TestListedBuildingStandardisationProcess(SparkTestCase):
             "listed_building_outline_data": _listed_building_outline_multirow_raw_df(spark),
         }
 
-        with mock.patch("odw.core.etl.transformation.standardised.listed_building_standardisation_process.LoggingUtil"):
-            inst = ListedBuildingStandardisationProcess(spark)
-            data_to_write, _ = inst.process(source_data=source_data)
+        inst = ListedBuildingStandardisationProcess(spark)
+        data_to_write, _ = inst.process(source_data=source_data)
 
         listed_building_df = data_to_write[LISTED_BUILDING_OUTPUT_TABLE]["data"]
         listed_building_outline_df = data_to_write[LISTED_BUILDING_OUTLINE_OUTPUT_TABLE]["data"]
@@ -671,9 +664,8 @@ class TestListedBuildingStandardisationProcess(SparkTestCase):
             "listed_building_outline_data": _listed_building_outline_empty_entities_raw_df(spark),
         }
 
-        with mock.patch("odw.core.etl.transformation.standardised.listed_building_standardisation_process.LoggingUtil"):
-            inst = ListedBuildingStandardisationProcess(spark)
-            data_to_write, _ = inst.process(source_data=source_data)
+        inst = ListedBuildingStandardisationProcess(spark)
+        data_to_write, _ = inst.process(source_data=source_data)
 
         listed_building_df = data_to_write[LISTED_BUILDING_OUTPUT_TABLE]["data"]
         listed_building_outline_df = data_to_write[LISTED_BUILDING_OUTLINE_OUTPUT_TABLE]["data"]
@@ -691,9 +683,8 @@ class TestListedBuildingStandardisationProcess(SparkTestCase):
             "listed_building_outline_data": _empty_listed_building_outline_raw_df(spark),
         }
 
-        with mock.patch("odw.core.etl.transformation.standardised.listed_building_standardisation_process.LoggingUtil"):
-            inst = ListedBuildingStandardisationProcess(spark)
-            data_to_write, _ = inst.process(source_data=source_data)
+        inst = ListedBuildingStandardisationProcess(spark)
+        data_to_write, _ = inst.process(source_data=source_data)
 
         listed_building_df = data_to_write[LISTED_BUILDING_OUTPUT_TABLE]["data"]
         listed_building_outline_df = data_to_write[LISTED_BUILDING_OUTLINE_OUTPUT_TABLE]["data"]
@@ -711,9 +702,8 @@ class TestListedBuildingStandardisationProcess(SparkTestCase):
             "listed_building_outline_data": _empty_listed_building_outline_raw_df(spark),
         }
 
-        with mock.patch("odw.core.etl.transformation.standardised.listed_building_standardisation_process.LoggingUtil"):
-            inst = ListedBuildingStandardisationProcess(spark)
-            data_to_write, _ = inst.process(source_data=source_data)
+        inst = ListedBuildingStandardisationProcess(spark)
+        data_to_write, _ = inst.process(source_data=source_data)
 
         listed_building_df = data_to_write[LISTED_BUILDING_OUTPUT_TABLE]["data"]
         listed_building_outline_df = data_to_write[LISTED_BUILDING_OUTLINE_OUTPUT_TABLE]["data"]
@@ -731,10 +721,289 @@ class TestListedBuildingStandardisationProcess(SparkTestCase):
             "listed_building_outline_data": _empty_listed_building_outline_raw_df(spark),
         }
 
-        with mock.patch("odw.core.etl.transformation.standardised.listed_building_standardisation_process.LoggingUtil"):
-            inst = ListedBuildingStandardisationProcess(spark)
-            data_to_write, _ = inst.process(source_data=source_data)
+        inst = ListedBuildingStandardisationProcess(spark)
+        data_to_write, _ = inst.process(source_data=source_data)
 
+        listed_building_df = data_to_write[LISTED_BUILDING_OUTPUT_TABLE]["data"]
+
+        assert listed_building_df.count() == 2
+        assert listed_building_df.where("entity = '1001'").count() == 2
+
+    def test__listed_building_standardisation_process__run__writes_both_outputs_end_to_end_like_legacy(
+        self,
+    ):
+        spark = PytestSparkSessionUtil().get_spark_session()
+
+        source_data = {
+            "listed_building_data": _listed_building_raw_df(spark),
+            "listed_building_outline_data": _empty_listed_building_outline_raw_df(spark),
+        }
+
+        inst = ListedBuildingStandardisationProcess(spark)
+
+        with (
+            mock.patch.object(inst, "load_data", return_value=source_data),
+            mock.patch.object(inst, "write_data") as mock_write,
+        ):
+            inst.run()
+
+        data_to_write = mock_write.call_args[0][0]
+
+        assert LISTED_BUILDING_OUTPUT_TABLE in data_to_write
+        assert LISTED_BUILDING_OUTLINE_OUTPUT_TABLE in data_to_write
+
+        listed_building_df = data_to_write[LISTED_BUILDING_OUTPUT_TABLE]["data"]
+        listed_building_outline_df = data_to_write[LISTED_BUILDING_OUTLINE_OUTPUT_TABLE]["data"]
+
+        assert listed_building_df.count() == 2
+        assert listed_building_outline_df.count() == 2
+
+        assert data_to_write[LISTED_BUILDING_OUTPUT_TABLE]["write_mode"] == "overwrite"
+        assert data_to_write[LISTED_BUILDING_OUTLINE_OUTPUT_TABLE]["write_mode"] == "overwrite"
+
+        assert listed_building_df.columns == [
+            "dataset",
+            "end-date",
+            "entity",
+            "entry-date",
+            "geometry",
+            "name",
+            "organisation-entity",
+            "point",
+            "prefix",
+            "reference",
+            "start-date",
+            "typology",
+            "documentation-url",
+            "listed-building-grade",
+        ]
+
+        assert listed_building_outline_df.columns == [
+            "address",
+            "address-text",
+            "dataset",
+            "document-url",
+            "documentation-url",
+            "end-date",
+            "entity",
+            "entry-date",
+            "geometry",
+            "listed-building",
+            "name",
+            "notes",
+            "organisation-entity",
+            "point",
+            "prefix",
+            "reference",
+            "start-date",
+            "typology",
+        ]
+
+    def test__listed_building_standardisation_process__run__preserves_flattened_values_and_nulls_like_legacy(
+        self,
+    ):
+        spark = PytestSparkSessionUtil().get_spark_session()
+
+        source_data = {
+            "listed_building_data": _listed_building_raw_df(spark),
+            "listed_building_outline_data": _empty_listed_building_outline_raw_df(spark),
+        }
+
+        inst = ListedBuildingStandardisationProcess(spark)
+
+        with (
+            mock.patch.object(inst, "load_data", return_value=source_data),
+            mock.patch.object(inst, "write_data") as mock_write,
+        ):
+            inst.run()
+
+        data_to_write = mock_write.call_args[0][0]
+
+        listed_building_df = data_to_write[LISTED_BUILDING_OUTPUT_TABLE]["data"]
+        listed_building_outline_df = data_to_write[LISTED_BUILDING_OUTLINE_OUTPUT_TABLE]["data"]
+
+        lb_row = (
+            listed_building_df.where("entity = '1001'")
+            .select(
+                "name",
+                "reference",
+                "listed-building-grade",
+                "geometry",
+                "documentation-url",
+            )
+            .collect()[0]
+        )
+        lb_null_row = listed_building_df.where("entity = '1002'").select("geometry", "point", "documentation-url").collect()[0]
+        lbo_row = (
+            listed_building_outline_df.where("entity = '2001'")
+            .select("address", "listed-building", "reference", "document-url", "notes")
+            .collect()[0]
+        )
+        lbo_null_row = (
+            listed_building_outline_df.where("entity = '2002'")
+            .select(
+                "address",
+                "address-text",
+                "document-url",
+                "documentation-url",
+                "geometry",
+                "notes",
+                "point",
+            )
+            .collect()[0]
+        )
+
+        assert lb_row["name"] == "Building One"
+        assert lb_row["reference"] == "LB-001"
+        assert lb_row["listed-building-grade"] == "II"
+        assert lb_row["geometry"] == "POLYGON((1 1,2 2,3 3,1 1))"
+        assert lb_row["documentation-url"] == "https://example.com/lb-001"
+
+        assert lb_null_row["geometry"] is None
+        assert lb_null_row["point"] is None
+        assert lb_null_row["documentation-url"] is None
+
+        assert lbo_row["address"] == "1 High Street"
+        assert lbo_row["listed-building"] == "1001"
+        assert lbo_row["reference"] == "LBO-001"
+        assert lbo_row["document-url"] == "https://example.com/doc-1"
+        assert lbo_row["notes"] == "Some notes"
+
+        assert lbo_null_row["address"] is None
+        assert lbo_null_row["address-text"] is None
+        assert lbo_null_row["document-url"] is None
+        assert lbo_null_row["documentation-url"] is None
+        assert lbo_null_row["geometry"] is None
+        assert lbo_null_row["notes"] is None
+        assert lbo_null_row["point"] is None
+
+    def test__listed_building_standardisation_process__run__explodes_entities_from_multiple_raw_rows(
+        self,
+    ):
+        spark = PytestSparkSessionUtil().get_spark_session()
+
+        source_data = {
+            "listed_building_data": _listed_building_multirow_raw_df(spark),
+            "listed_building_outline_data": _listed_building_outline_multirow_raw_df(spark),
+        }
+
+        inst = ListedBuildingStandardisationProcess(spark)
+
+        with (
+            mock.patch.object(inst, "load_data", return_value=source_data),
+            mock.patch.object(inst, "write_data") as mock_write,
+        ):
+            inst.run()
+
+        data_to_write = mock_write.call_args[0][0]
+        listed_building_df = data_to_write[LISTED_BUILDING_OUTPUT_TABLE]["data"]
+        listed_building_outline_df = data_to_write[LISTED_BUILDING_OUTLINE_OUTPUT_TABLE]["data"]
+
+        assert listed_building_df.count() == 3
+        assert listed_building_outline_df.count() == 3
+
+        listed_building_entities = {row["entity"] for row in listed_building_df.select("entity").collect()}
+        listed_building_outline_entities = {row["entity"] for row in listed_building_outline_df.select("entity").collect()}
+
+        assert listed_building_entities == {"1001", "1002", "1003"}
+        assert listed_building_outline_entities == {"2001", "2002", "2003"}
+
+    def test__listed_building_standardisation_process__run__empty_entities_arrays_return_empty_outputs(
+        self,
+    ):
+        spark = PytestSparkSessionUtil().get_spark_session()
+
+        source_data = {
+            "listed_building_data": _listed_building_empty_entities_raw_df(spark),
+            "listed_building_outline_data": _listed_building_outline_empty_entities_raw_df(spark),
+        }
+
+        inst = ListedBuildingStandardisationProcess(spark)
+
+        with (
+            mock.patch.object(inst, "load_data", return_value=source_data),
+            mock.patch.object(inst, "write_data") as mock_write,
+        ):
+            inst.run()
+
+        data_to_write = mock_write.call_args[0][0]
+        listed_building_df = data_to_write[LISTED_BUILDING_OUTPUT_TABLE]["data"]
+        listed_building_outline_df = data_to_write[LISTED_BUILDING_OUTLINE_OUTPUT_TABLE]["data"]
+
+        assert listed_building_df.count() == 0
+        assert listed_building_outline_df.count() == 0
+
+    def test__listed_building_standardisation_process__run__empty_input_dataframes_return_empty_outputs(
+        self,
+    ):
+        spark = PytestSparkSessionUtil().get_spark_session()
+
+        source_data = {
+            "listed_building_data": _empty_listed_building_raw_df(spark),
+            "listed_building_outline_data": _empty_listed_building_outline_raw_df(spark),
+        }
+
+        inst = ListedBuildingStandardisationProcess(spark)
+
+        with (
+            mock.patch.object(inst, "load_data", return_value=source_data),
+            mock.patch.object(inst, "write_data") as mock_write,
+        ):
+            inst.run()
+
+        data_to_write = mock_write.call_args[0][0]
+        listed_building_df = data_to_write[LISTED_BUILDING_OUTPUT_TABLE]["data"]
+        listed_building_outline_df = data_to_write[LISTED_BUILDING_OUTLINE_OUTPUT_TABLE]["data"]
+
+        assert listed_building_df.count() == 0
+        assert listed_building_outline_df.count() == 0
+        assert data_to_write[LISTED_BUILDING_OUTPUT_TABLE]["write_mode"] == "overwrite"
+        assert data_to_write[LISTED_BUILDING_OUTLINE_OUTPUT_TABLE]["write_mode"] == "overwrite"
+
+    def test__listed_building_standardisation_process__run__one_source_empty_still_writes_the_other_output(
+        self,
+    ):
+        spark = PytestSparkSessionUtil().get_spark_session()
+
+        source_data = {
+            "listed_building_data": _empty_listed_building_raw_df(spark),
+            "listed_building_outline_data": _empty_listed_building_outline_raw_df(spark),
+        }
+
+        inst = ListedBuildingStandardisationProcess(spark)
+
+        with (
+            mock.patch.object(inst, "load_data", return_value=source_data),
+            mock.patch.object(inst, "write_data") as mock_write,
+        ):
+            inst.run()
+
+        data_to_write = mock_write.call_args[0][0]
+        listed_building_df = data_to_write[LISTED_BUILDING_OUTPUT_TABLE]["data"]
+        listed_building_outline_df = data_to_write[LISTED_BUILDING_OUTLINE_OUTPUT_TABLE]["data"]
+
+        assert listed_building_df.count() == 2
+        assert listed_building_outline_df.count() == 0
+
+    def test__listed_building_standardisation_process__run__preserves_duplicate_entities_like_legacy(
+        self,
+    ):
+        spark = PytestSparkSessionUtil().get_spark_session()
+
+        source_data = {
+            "listed_building_data": _duplicate_listed_building_raw_df(spark),
+            "listed_building_outline_data": _empty_listed_building_outline_raw_df(spark),
+        }
+
+        inst = ListedBuildingStandardisationProcess(spark)
+
+        with (
+            mock.patch.object(inst, "load_data", return_value=source_data),
+            mock.patch.object(inst, "write_data") as mock_write,
+        ):
+            inst.run()
+
+        data_to_write = mock_write.call_args[0][0]
         listed_building_df = data_to_write[LISTED_BUILDING_OUTPUT_TABLE]["data"]
 
         assert listed_building_df.count() == 2
